+++ lib/generators/test_unit/webauthn_authentication/templates/test/system/manage_webauthn_credentials_test.rb
@@ -5,8 +5,7 @@ class ManageWebauthnCredentialsTest < ApplicationSystemTestCase
   include VirtualAuthenticatorTestHelper
 
   def setup
-    user = User.create!(email_address: "alice@example.com", password: "S3cr3tP@ssw0rd!")
-    sign_in_as(user)
+    @user = User.create!(email_address: "alice@example.com", password: "S3cr3tP@ssw0rd!")
     @authenticator = add_virtual_authenticator
   end
 
@@ -14,63 +13,64 @@ class ManageWebauthnCredentialsTest < ApplicationSystemTestCase
     @authenticator.remove!
   end
 
-  test "add credentials and sign in" do
-    visit root_path
+  test "adding a passkey" do
+    sign_in_as(@user)
 
-    click_on "Add Passkey"
-
+    visit new_passkey_path
     fill_in("Passkey nickname", with: "Touch ID")
     click_on "Add Passkey"
 
-    assert_current_path "/"
-    assert_selector "div", text: "Passkey registered successfully"
-    assert_selector "span", text: "Touch ID"
-
-    click_on "Sign out"
-    assert_selector("input[type=submit][value='Sign in']")
-
-    click_on "Sign In with Passkey"
-
-    assert_current_path "/"
-    assert_selector "h3", text: "Your Passkeys"
+    assert_current_path root_path
+    # Add custom assertions based on your application's behavior
+    # assert_text "Passkey registered successfully"
   end
 
-  test "sign in with 2FA WebAuthn credential" do
-    visit root_path
+  test "signing in with existing passkey" do
+    add_passkey_to_authenticator(@authenticator, @user)
 
-    click_on "Add Second Factor Key"
+    visit new_session_path
+    click_on "Sign In with Passkey"
 
+    assert_current_path root_path
+    # Add custom assertions based on your application's behavior
+  end
+
+  test "adding a 2FA WebAuthn credential" do
+    sign_in_as(@user)
+
+    visit new_second_factor_webauthn_credential_path
     fill_in("Security Key nickname", with: "Touch ID")
     click_on "Add Security Key"
 
-    assert_current_path "/"
-    assert_selector "div", text: "Security Key registered successfully"
-    assert_selector "span", text: "Touch ID"
+    assert_current_path root_path
+    # Add custom assertions based on your application's behavior
+    # assert_text "Security Key registered successfully"
+  end
 
-    click_on "Sign out"
-    assert_selector("input[type=submit][value='Sign in']")
+  test "sign in with existing 2FA WebAuthn credential" do
+    add_security_key_to_authenticator(@authenticator, @user)
 
-    fill_in "email_address", with: "alice@example.com"
-    fill_in "password", with: "S3cr3tP@ssw0rd!"
+    visit new_session_path
+    fill_in "email_address", with: @user.email_address
+    fill_in "password", with: @user.password
     click_on "Sign in"
 
+    assert_current_path new_second_factor_authentication_path
     assert_selector "h3", text: "Two-factor authentication"
     click_on "Use Security Key"
 
-    assert_current_path "/"
-    assert_selector "h3", text: "Your Passkeys"
+    assert_current_path root_path
+    # Add custom assertions based on your application's behavior
   end
 
   private
 
   def sign_in_as(user)
     visit new_session_path
-
     fill_in "email_address", with: user.email_address
     fill_in "password", with: user.password
-
     click_on "Sign in"
 
-    assert_selector "h3", text: "Your Passkeys"
+    assert_current_path root_path
   end
 end